\name{eigentest-package}
\alias{eigentest-package}
\alias{eigentest}
\docType{package}
\title{
\packageTitle{eigentest}
}
\description{
\packageDescription{eigentest}
}
\details{
\tabular{ll}{
Package: \tab eigentest\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2020-05-20\cr
License: \tab GPL-2\cr
}
}

\keyword{ package }

